More updates to slide
Added all stuff to the new scene
Reworked the cart. the player now is able to ride it but are put in place. They can still look around.
Added checkpoint
Working on a goal for the game for the player to bring a rock, to the bottom of the slope.